# contracts/openapi_v0.yaml
openapi: 3.0.3
info:
  title: AI Language Partner Mobile API
  version: 0.1.0
  description: Shared API contract for Claude frontend and Codex backend. Do not change without contract request.
servers:
  - url: http://localhost:8000
paths:
...
  /v1/content/import:
    post:
      operationId: importContentBundle
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentImportRequest' }
      responses:
        '200':
          description: content import dry-run or applied result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  dryRun: { type: boolean }
                  applied: { type: boolean }
                  importedCounts:
                    type: object
                    properties:
                      courses: { type: integer }
                      practiceRooms: { type: integer }
                  version: { $ref: '#/components/schemas/ContentVersionSummary' }
                  report: { $ref: '#/components/schemas/ContentQualityReport' }
...
  /v1/content/versions/{versionId}/submit-review:
    post:
      operationId: submitContentVersionReview
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - $ref: '#/components/parameters/AdminRoleHeader'
        - $ref: '#/components/parameters/AdminUserHeader'
        - in: path
          name: versionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentReviewRequest' }
      responses:
        '200':
          description: content version submitted for reviewer approval
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentVersionWorkflowResponse' }
  /v1/content/versions/{versionId}/approve:
    post:
      operationId: approveContentVersion
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - $ref: '#/components/parameters/AdminRoleHeader'
        - $ref: '#/components/parameters/AdminUserHeader'
        - in: path
          name: versionId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentReviewRequest' }
      responses:
        '200':
          description: content version approved by a reviewer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ContentVersionWorkflowResponse' }
...
  /v1/content/operations/jobs:
    get:
      operationId: listContentOperationJobs
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - $ref: '#/components/parameters/AdminRoleHeader'
        - $ref: '#/components/parameters/AdminUserHeader'
        - in: query
          name: status
          required: false
          schema: { type: string, enum: [queued, running, succeeded, failed, canceled] }
        - in: query
          name: limit
          required: false
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
      responses:
        '200':
          description: queued and historical content operation jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items: { $ref: '#/components/schemas/ContentOperationJob' }
    post:
      operationId: createContentOperationJob
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - $ref: '#/components/parameters/AdminRoleHeader'
        - $ref: '#/components/parameters/AdminUserHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentOperationJobRequest' }
      responses:
        '200':
          description: queued a content operation job
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  job: { $ref: '#/components/schemas/ContentOperationJob' }
...
  /v1/content/scheduler/runs:
    get:
      operationId: listContentSchedulerRuns
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - $ref: '#/components/parameters/AdminRoleHeader'
        - $ref: '#/components/parameters/AdminUserHeader'
        - in: query
          name: status
          required: false
          schema: { type: string, enum: [running, succeeded, failed] }
        - in: query
          name: limit
          required: false
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
      responses:
        '200':
          description: managed content scheduler run history
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items: { $ref: '#/components/schemas/ContentSchedulerRun' }
  /v1/content/scheduler/run-once:
    post:
      operationId: runContentSchedulerOnce
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - $ref: '#/components/parameters/AdminRoleHeader'
        - $ref: '#/components/parameters/AdminUserHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentSchedulerRunOnceRequest' }
      responses:
        '200':
          description: ran one managed content scheduler tick across due releases and queued operation jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  run: { $ref: '#/components/schemas/ContentSchedulerRun' }
                  releaseWorker:
                    type: object
                    additionalProperties: true
                  operationJobs:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
  /v1/content/releases:
    get:
      operationId: listContentReleases
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - $ref: '#/components/parameters/AdminRoleHeader'
        - $ref: '#/components/parameters/AdminUserHeader'
        - in: query
          name: status
          required: false
          schema: { type: string, enum: [planned, scheduled, applied, rolled_back, canceled] }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: content release plans and release history
          content:
            application/json:
              schema:
                type: object
                properties:
                  releases:
                    type: array
                    items: { $ref: '#/components/schemas/ContentRelease' }
    post:
      operationId: createContentRelease
      parameters:
        - $ref: '#/components/parameters/AdminKeyHeader'
        - $ref: '#/components/parameters/AdminRoleHeader'
        - $ref: '#/components/parameters/AdminUserHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContentReleaseRequest' }
      responses:
        '200':
          description: planned content release with quality guardrail snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  release: { $ref: '#/components/schemas/ContentRelease' }
                  report: { $ref: '#/components/schemas/ContentQualityReport' }
...
components:
  parameters:
...
    AdminKeyHeader:
      in: header
      name: X-Admin-Key
      required: true
      schema: { type: string }
      description: Admin key required for audit-log access.
    AdminRoleHeader:
      in: header
      name: X-Admin-Role
      required: false
      schema: { type: string, enum: [viewer, editor, reviewer, publisher, owner] }
      description: Optional admin role for content workflow gates. Missing role defaults to owner after a valid admin key.
    AdminUserHeader:
      in: header
      name: X-Admin-User
      required: false
      schema: { type: string }
      description: Optional admin actor label stored in content workflow audit metadata.
...
  schemas:
...
    PracticeRoom:
      type: object
      required: [id, title, primaryPhraseKo, primaryPhraseJa, personaId]
      properties:
        id: { type: string }
        title: { type: string }
        primaryPhraseKo: { type: string }
        primaryPhraseJa: { type: string }
        alternativePhrasesJa:
          type: array
          items: { type: string }
        personaId: { type: string }
        scenario: { type: string }
        openingMessage: { type: string }
        courseId: { type: string }
        courseTitle: { type: string }
        unitId: { type: string }
        unitTitle: { type: string }
        unitOrder: { type: integer }
        lessonId: { type: string }
        lessonTitle: { type: string }
        lessonOrder: { type: integer }
        roomOrder: { type: integer }
        tags:
          type: array
          items: { type: string }
    Course:
      type: object
      required: [id, title, targetLanguage, nativeLanguage, level, units]
      properties:
        id: { type: string }
        title: { type: string }
        targetLanguage: { type: string }
        nativeLanguage: { type: string }
        level: { type: string }
        descriptionKo: { type: string }
        units:
          type: array
          items: { $ref: '#/components/schemas/CourseUnit' }
    CourseUnit:
      type: object
      required: [id, title, order, lessons]
      properties:
        id: { type: string }
        title: { type: string }
        order: { type: integer }
        skillTags:
          type: array
          items: { type: string }
        lessons:
          type: array
          items: { $ref: '#/components/schemas/CourseLesson' }
    CourseLesson:
      type: object
      required: [id, title, order, practiceRoomIds]
      properties:
        id: { type: string }
        title: { type: string }
        order: { type: integer }
        practiceRoomIds:
          type: array
          items: { type: string }
    ContentBundleRequest:
      type: object
      properties:
        courses:
          type: array
          items: { $ref: '#/components/schemas/Course' }
        practiceRooms:
          type: array
          items: { $ref: '#/components/schemas/PracticeRoom' }
    ContentImportRequest:
      allOf:
        - $ref: '#/components/schemas/ContentBundleRequest'
        - type: object
          properties:
            dryRun: { type: boolean, default: true }
            replaceExisting: { type: boolean, default: true }
    ContentImportedCounts:
      type: object
      properties:
        courses: { type: integer }
        practiceRooms: { type: integer }
    ContentSnapshotCounts:
      type: object
      properties:
        courses: { type: integer }
        practiceRooms: { type: integer }
    ContentVersionSummary:
      type: object
      required: [id, status, source, createdAt, importedCounts, report, snapshotCounts]
      properties:
        id: { type: string }
        label: { type: string, nullable: true }
        status: { type: string, enum: [draft, in_review, approved, rejected, published] }
        parentVersionId: { type: string, nullable: true }
        branchName: { type: string, nullable: true }
        source: { type: string }
        createdBy: { type: string, nullable: true }
        submittedBy: { type: string, nullable: true }
        reviewedBy: { type: string, nullable: true }
        reviewNote: { type: string, nullable: true }
        createdAt: { type: string }
        submittedAt: { type: string, nullable: true }
        reviewedAt: { type: string, nullable: true }
        publishedAt: { type: string, nullable: true }
        importedCounts: { $ref: '#/components/schemas/ContentImportedCounts' }
        report: { $ref: '#/components/schemas/ContentQualityReport' }
        snapshotCounts: { $ref: '#/components/schemas/ContentSnapshotCounts' }
    ContentReviewRequest:
      type: object
      properties:
        note: { type: string, nullable: true }
...
    ContentReleaseRequest:
      type: object
      required: [versionId, title]
      properties:
        versionId: { type: string, minLength: 1, maxLength: 120 }
        title: { type: string, minLength: 1, maxLength: 160 }
        releaseStrategy: { type: string, enum: [immediate, scheduled, canary], default: immediate }
        rolloutPercent: { type: integer, minimum: 1, maximum: 100, default: 100 }
        catalogScope: { type: string, enum: [incremental, full_catalog], default: incremental }
        scheduledAt: { type: string, nullable: true }
        guardrails:
          type: object
          additionalProperties: true
        note: { type: string, maxLength: 1000, nullable: true }
...
    ContentOperationJobRequest:
      type: object
      required: [jobType]
      properties:
        jobType: { type: string, enum: [validate_bundle, import_bundle, run_due_releases] }
        priority: { type: string, enum: [low, normal, high, urgent], default: normal }
        payload:
          type: object
          additionalProperties: true
...
    ContentSchedulerRunOnceRequest:
      type: object
      required: [confirmation]
      properties:
        confirmation: { type: string, enum: [run-content-scheduler-once] }
        schedulerKey: { type: string, minLength: 1, maxLength: 80, default: content_ops }
        leaseOwner: { type: string, minLength: 1, maxLength: 120, default: local-scheduler }
        maxOperationJobs: { type: integer, minimum: 0, maximum: 20, default: 1 }
        releaseLimit: { type: integer, minimum: 1, maximum: 200, default: 50 }
    ContentOperationJob:
      type: object
      required: [id, jobType, status, priority, payload, result, createdAt, updatedAt]
      properties:
        id: { type: string }
        jobType: { type: string, enum: [validate_bundle, import_bundle, run_due_releases] }
        status: { type: string, enum: [queued, running, succeeded, failed, canceled] }
        priority: { type: string, enum: [low, normal, high, urgent] }
        payload:
          type: object
          additionalProperties: true
        result:
          type: object
          additionalProperties: true
        error: { type: string, nullable: true }
        createdBy: { type: string, nullable: true }
        claimedBy: { type: string, nullable: true }
        canceledBy: { type: string, nullable: true }
        createdAt: { type: string }
        updatedAt: { type: string }
        claimedAt: { type: string, nullable: true }
        completedAt: { type: string, nullable: true }
        canceledAt: { type: string, nullable: true }
    ContentSchedulerRun:
      type: object
      required: [id, schedulerKey, status, leaseOwner, startedAt, heartbeatAt, maxOperationJobs, releaseLimit, result]
      properties:
        id: { type: string }
        schedulerKey: { type: string }
        status: { type: string, enum: [running, succeeded, failed] }
        leaseOwner: { type: string }
        actor: { type: string, nullable: true }
        startedAt: { type: string }
        heartbeatAt: { type: string }
        completedAt: { type: string, nullable: true }
        maxOperationJobs: { type: integer }
        releaseLimit: { type: integer }
        result:
          type: object
          additionalProperties: true
        error: { type: string, nullable: true }
    ContentRelease:
      type: object
      required: [id, versionId, title, status, releaseStrategy, rolloutPercent, catalogScope, guardrails, createdAt]
      properties:
        id: { type: string }
        versionId: { type: string }
        title: { type: string }
        status: { type: string, enum: [planned, scheduled, applied, rolled_back, canceled] }
        releaseStrategy: { type: string, enum: [immediate, scheduled, canary] }
        rolloutPercent: { type: integer }
        catalogScope: { type: string, enum: [incremental, full_catalog] }
        scheduledAt: { type: string, nullable: true }
        guardrails:
          type: object
          additionalProperties: true
        note: { type: string, nullable: true }
        previousPublishedVersionId: { type: string, nullable: true }
        importedCounts: { $ref: '#/components/schemas/ContentImportedCounts' }
        rollbackImportedCounts: { $ref: '#/components/schemas/ContentImportedCounts' }
        createdBy: { type: string, nullable: true }
        appliedBy: { type: string, nullable: true }
        rolledBackBy: { type: string, nullable: true }
        createdAt: { type: string }
        appliedAt: { type: string, nullable: true }
        rolledBackAt: { type: string, nullable: true }
        rollbackNote: { type: string, nullable: true }
    ContentVersionWorkflowResponse:
      type: object
      properties:
        ok: { type: boolean }
        version: { $ref: '#/components/schemas/ContentVersion' }
    ContentVersion:
      allOf:
        - $ref: '#/components/schemas/ContentVersionSummary'
        - type: object
          required: [courses, practiceRooms]
          properties:
            courses:
              type: array
              items: { $ref: '#/components/schemas/Course' }
            practiceRooms:
              type: array
              items: { $ref: '#/components/schemas/PracticeRoom' }
    ContentQualityIssue:
      type: object
      required: [code, message, pointer]
      properties:
        code: { type: string }
        message: { type: string }
        pointer: { type: string }
      additionalProperties: true
    ContentQualityReport:
      type: object
      required: [valid, errors, warnings, counts]
      properties:
        valid: { type: boolean }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ContentQualityIssue' }
        warnings:
          type: array
          items: { $ref: '#/components/schemas/ContentQualityIssue' }
        counts:
          type: object
          properties:
            personas: { type: integer }
            courses: { type: integer }
            units: { type: integer }
            lessons: { type: integer }
            practiceRooms: { type: integer }
            practiceRoomRefs: { type: integer }
            roomsWithCoursePlacement: { type: integer }
        courseIds:
          type: array
          items: { type: string }
        missingPersonaIds:
          type: array
          items: { type: string }
        missingPracticeRoomRefs:
          type: array
          items: { type: string }
        duplicatePracticeRoomRefs:
          type: array
          items: { type: string }
        orphanPracticeRoomIds:
          type: array
          items: { type: string }
        referencedPracticeRoomIds:
          type: array
          items: { type: string }
...